byte -> int for EntityType

+++ DIMP/src/main/java/chat/dim/mkm/Entity.java
@@ -59,7 +59,7 @@ public interface Entity {
      *
      * @return network type
      */
-    byte getType();
+    int getType();
 
     void setDataSource(DataSource dataSource);
     DataSource getDataSource();
